requirements.txt
Flask==2.3.3
spotipy==2.23.0
python-dotenv==1.0.0
tqdm==4.66.1